+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
 * Initial Release
+
+## 0.0.2
+
+* Fix analysis
+* Random Verse
+* Search Chapters or Verses